refactor: change package names and get rid of default SW360 dependency
- Do not use SW360 Report Generator to get rid of SW360 Dependency
- rename some packages from util => jsonReader to get rid of signing problems
