+++ antenna-basic-assembly/pom.xml
@@ -31,7 +31,6 @@
         <!-- knowledge base -->
         <module>antenna-spdx-license-knowledge-base</module>
         <!-- workflow steps -->
-        <module>generators/antenna-sw360-disclosure-document-generator</module>
         <module>antenna-basic-rules</module>
     </modules>
