+++ antenna-documentation/src/site/markdown/generators/generators.md
@@ -5,5 +5,4 @@ Antenna already provides the five following generators:
 - [HTML generator](./HTML-report-generator-step.html)
 - [CSV generator](./csv-generator-step.html)
 - [Source Zip Generator](./source-zip-generator-step.html)
-- [SW360 report generator](./sw360-disclosure-document-generator-step.html)
 - [SW360 update generator](./sw360-update-generator-step.html)
