+++ antenna-documentation/src/site/markdown/outputHandlers/add-to-archive-handler.md
@@ -11,7 +11,7 @@ To add the output handler, add the following to the workflow.xml:
             <name>Add document to zip</name>
             <classHint>org.eclipse.sw360.antenna.workflow.outputHandlers.FileToArchiveWriter</classHint>
             <configuration>
-                <entry key="instructions" value="disclosure-sw360-doc-txt:${project.build.directory}/${project.artifactId}-${project.version}.jar:/legalnotice/DisclosureDoc.txt"/>
+                <entry key="instructions" value="disclosure-doc:${project.build.directory}/${project.artifactId}-${project.version}.jar:/legalnotice/3rdparty-licenses.html"/>
             </configuration>
         </step>
     </outputHandlers>
@@ -31,9 +31,9 @@ The magic strings may not contain additional colons.
 
 As an example, consider the string in the How to use section above. It consists of:
 
-- `disclosure-sw360-doc-txt`: The magic string for a file produced by the [SW360 disclosure document generator](../generators/sw360-disclosure-document-generator-step.html)
+- `disclosure-doc`: The magic string for a file produced by the [HTML Report Writer](../generators/HTML-report-generator-step.html)
 - `${project.build.directory}/${project.artifactId}-${project.version}.jar`
-- `/legalnotice/DisclosureDoc.txt`
+- `/legalnotice/3rdparty-licenses.html`
 
 This will result in the output file generated by the SW360 disclosure document generator to be added to jar of the artifact being built.
 The file will then reside in `/legalnotice/DisclosureDoc.txt` inside the project jar.
