+++ antenna-testing/antenna-frontend-stubs-testing/src/main/java/org/eclipse/sw360/antenna/frontend/testProjects/P2TestProject.java
@@ -155,7 +155,7 @@ public class P2TestProject extends AbstractTestProjectWithExpectations implement
 
     @Override
     public Collection<String> getExpectedBuildArtifacts() {
-        return Stream.of("artifact-information", "disclosure-doc", "disclosure-sw360-doc-txt").collect(Collectors.toSet());
+        return Stream.of("artifact-information", "disclosure-doc").collect(Collectors.toSet());
     }
 
     @Override
