+++ antenna-workflow-steps/analyzers/antenna-json-analyzer/src/test/java/org/eclipse/sw360/antenna/jsonreader/JsonReaderTest.java
@@ -8,7 +8,7 @@
  *
  * SPDX-License-Identifier: EPL-2.0
  */
-package org.eclipse.sw360.antenna.util;
+package org.eclipse.sw360.antenna.jsonreader;
 
 import org.eclipse.sw360.antenna.jsonreader.JsonReader;
 import org.eclipse.sw360.antenna.model.artifact.Artifact;
