+++ example-projects/mvn-test-project/src/workflow.xml
@@ -44,20 +44,13 @@
             </configuration>
             <deactivated>true</deactivated>
         </step>
-        <step>
-            <name>SW360 Report Generator</name>
-            <classHint>org.eclipse.sw360.antenna.workflow.generators.SW360DisclosureDocumentGenerator</classHint>
-            <configuration>
-                <entry key="disclosure.doc.formats" value="txt"/>
-            </configuration>
-        </step>
     </generators>
     <outputHandlers>
         <step>
             <name>Add disclosure document to jar</name>
             <classHint>org.eclipse.sw360.antenna.workflow.outputHandlers.FileToArchiveWriter</classHint>
             <configuration>
-                <entry key="instructions" value="disclosure-sw360-doc-txt:${project.build.directory}/${project.artifactId}-${project.version}.jar:/legalnotice/DisclosureDoc.txt"/>
+                <entry key="instructions" value="disclosure-doc:${project.build.directory}/${project.artifactId}-${project.version}.jar:/legalnotice/DisclosureDoc.html"/>
             </configuration>
         </step>
     </outputHandlers>
